pyproject.toml
[project]
name = "sharemov"
version = "0.1.0"
description = "P2P Video Call, Chat & Screen Share over WebRTC"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "python-socketio>=5.12.0",
    "jinja2>=3.1.0",
]
